Sflow support for ONOS
sFlow is a technology for monitoring traffic in data networks
containing switches and routers.  In particular, it defines the
sampling mechanisms implemented in an sFlow Agent for monitoring
traffic, the sFlow MIB for controlling the sFlow Agent, and the
format of sample data used by the sFlow Agent when forwarding data to
a central data collector.

Rfc: rfc3176
Change-Id: I74ce40d86c0cea1defe5f8c0947d1512b5843eb5